feat(rust): add confirmation prompt to delete commands

+++ implementations/rust/ockam/ockam_command/src/identity/delete.rs
@@ -19,6 +19,10 @@ after_long_help = docs::after_help(AFTER_LONG_HELP)
 pub struct DeleteCommand {
     /// Name of the identity to be deleted
     name: String,
+
+    /// Confirm the deletion without prompting
+    #[arg(display_order = 901, long, short)]
+    yes: bool,
 }
 
 impl DeleteCommand {
@@ -33,15 +37,20 @@ async fn run_impl(
 ) -> miette::Result<()> {
     let state = opts.state;
     let idt = state.identities.get(&cmd.name)?;
-    state.delete_identity(idt)?;
-    opts.terminal
-        .stdout()
-        .plain(fmt_ok!(
-            "The identity named '{}' has been deleted.",
-            &cmd.name
-        ))
-        .machine(&cmd.name)
-        .json(serde_json::json!({ "name": &cmd.name }))
-        .write_line()?;
+    if opts
+        .terminal
+        .confirmed_with_flag_or_prompt(cmd.yes, "Are you sure you want to delete this identity?")?
+    {
+        state.delete_identity(idt)?;
+        opts.terminal
+            .stdout()
+            .plain(fmt_ok!(
+                "The identity named '{}' has been deleted",
+                &cmd.name
+            ))
+            .machine(&cmd.name)
+            .json(serde_json::json!({ "name": &cmd.name }))
+            .write_line()?;
+    }
     Ok(())
 }
